recombine ResourceDownloadHelpersQt with ResourceDownloadHelpers
