+++ SerialPrograms/Source/CommonFramework/ResourceDownload/ResourceDownloadHelpers.cpp
@@ -4,9 +4,13 @@
  *
  */
 
+#ifdef QT_CORE_LIB
+#include "CommonFramework/Tools/FileDownloader.h"
+#endif
 #include "Common/Cpp/Filesystem/Filesystem.h"
 #include "Common/Cpp/Json/JsonArray.h"
 #include "Common/Cpp/Json/JsonObject.h"
+#include "Common/Cpp/Json/JsonValue.h"
 #include "CommonFramework/GlobalAutoPaths.h"
 #include "CommonFramework/Logging/Logger.h"
 #include "CommonFramework/Exceptions/OperationFailedException.h"
@@ -161,5 +165,63 @@ const std::unordered_set<std::string>& all_resource_names(){
 
 
 }
+
+
+
+
+JsonValue fetch_resource_download_list_json_from_remote(){
+    Logger& logger = global_logger_tagged();
+#ifdef QT_CORE_LIB
+    JsonValue json = 
+        FileDownloader::download_json_file(
+            logger,
+            "https://raw.githubusercontent.com/PokemonAutomation/Packages/refs/heads/master/Resources/ResourceDownloadList.json"
+        );
+#else
+    JsonValue json; // placeholder when no Qt
+#endif
+    
+    return json;
+}
+
+
+const JsonValue& remote_resource_download_list_json(){
+    static const JsonValue json = fetch_resource_download_list_json_from_remote();
+
+    return json;
+}
+
+const std::vector<DownloadedResourceMetadata>& remote_resource_download_list(){
+    // cout << "remote_resource_download_list" << endl;
+    static std::vector<DownloadedResourceMetadata> remote_resources = deserialize_resource_list_json(remote_resource_download_list_json());
+
+    return remote_resources;
+}
+
+DownloadedResourceMetadata get_remote_resource_metadata_from_resource_slug(const std::string& target_resource_slug){
+    Logger& logger = global_logger_tagged();
+    std::vector<DownloadedResourceMetadata> remote_resources;
+
+    // Step 1: Attempt to fetch the list of available downloads
+    try{
+        remote_resources = remote_resource_download_list();
+    }catch(OperationFailedException&){
+        std::cerr << "get_remote_resource_metadata_from_resource_slug: Error" << endl;
+        throw_and_log<OperationFailedException>(logger, ErrorReport::NO_ERROR_REPORT, 
+            "Error: Download failed. Failed to fetch the list of available downloads. Check your internet connection.");
+    }
+
+    // Step 2: Attempt to extract metadata for the specific slug
+    try{
+        return get_resource_metadata_from_resource_type(target_resource_slug, remote_resources);
+    }catch(OperationFailedException&){
+        std::cerr << "get_remote_resource_metadata_from_resource_slug: Error" << endl;
+        throw_and_log<OperationFailedException>(logger, ErrorReport::NO_ERROR_REPORT, 
+            "get_remote_resource_metadata_from_resource_slug: Unable to find " + target_resource_slug + " within resource_list. "
+            "Likely caused by resource being no longer available for download. We recommend updating the Computer Control program.");
+    }
+}
+
+
 
 }
