+++ SerialPrograms/Source/CommonFramework/ResourceDownload/ResourceDownloadHelpers.h
@@ -72,5 +72,16 @@ DownloadedResourceMetadata get_expected_resource_metadata_from_resource_slug(con
 
 const std::unordered_set<std::string>& all_resource_names();
 
+
+
+const std::vector<DownloadedResourceMetadata>& remote_resource_download_list();
+
+
+// - throws OperationFailedException if target_resource_slug isn't found within remote_resource_download_list
+// this would indicate that CC is out of date.
+// - also throws OperationFailedException if Internet is not turned on.
+DownloadedResourceMetadata get_remote_resource_metadata_from_resource_slug(const std::string& target_resource_slug);
+
+
 }
 #endif
